+++ server/src/inmantals/pipeserver.py
@@ -21,6 +21,7 @@ from inmantals.server import InmantaLSHandler
 import logging
 from tornado.iostream import PipeIOStream
 import sys
+import os
 
 
 def main():
@@ -38,12 +39,12 @@ def main():
     stdin = PipeIOStream(sys.stdin.fileno())
     stdout = PipeIOStream(sys.stdout.fileno())
     handler = InmantaLSHandler(stdin, stdout, "0.0.0.0")
-    sys.stderr.write("starting")
+    sys.stderr.write(f"starting{os.linesep}")
     sys.stderr.flush()
 
     IOLoop.current().run_sync(handler.start)
 
-    sys.stderr.write("stopped")
+    sys.stderr.write(f"stopped{os.linesep}")
     sys.stderr.flush()