+++ src/test/index.ts
@@ -2,17 +2,19 @@ import * as path from 'path';
 import * as Mocha from 'mocha';
 import * as glob from 'glob';
 
-export function run(): Promise<void> {
+export function innerRun(regexToTestFiles: string): Promise<void> {
 	// Create the mocha test
 	const mocha = new Mocha({
 		ui: 'tdd',
 		color: true
 	});
 
-	const testsRoot = path.resolve(__dirname, '..');
+	const testsRoot = __dirname;
+	console.log(testsRoot)
 
 	return new Promise((c, e) => {
-		glob('**/**.test.js', { cwd: testsRoot }, (err, files) => {
+		glob(regexToTestFiles, { cwd: testsRoot }, (err, files) => {
+			console.log(files);
 			if (err) {
 				return e(err);
 			}