+++ src/test/navigation/navigation.test.ts
@@ -1,0 +1,70 @@
+import * as assert from 'assert';
+import { after, describe, it, beforeEach } from 'mocha';
+import * as path from 'path';
+import * as fs from 'fs-extra';
+
+import { Uri, window, commands, workspace, TextDocument, Position, Range, Location } from 'vscode';
+import { getInmantaVersion, waitForCompile } from '../helpers';
+
+const logPath: string = '/tmp/vscode-inmanta.log';
+const workspaceUri: Uri = Uri.file(path.resolve(__dirname, '../../../src/test/navigation/workspace'));
+const libsPath: string = path.resolve(workspaceUri.fsPath, 'libs');
+
+
+const modelUri: Uri = Uri.file(path.resolve(workspaceUri.fsPath, 'main.cf'));
+
+describe('Language Server Code navigation', () => {
+
+	beforeEach(async () => {
+		await Promise.all([
+			fs.writeFile(logPath, ""),
+		]).then(async values => {
+			await commands.executeCommand('workbench.action.closeActiveEditor');
+		});
+	});
+
+	
+	it(`Check that code navigation works`, () => {
+		return new Promise(async resolve => {			
+			// Open model file
+			const doc: TextDocument = await workspace.openTextDocument(modelUri);
+			await window.showTextDocument(doc);
+			const succeeded = await waitForCompile(logPath, 25000);
+			assert.strictEqual(succeeded, true, "Compilation didn't succeed");
+			const attributeInSameFile = await commands.executeCommand("vscode.executeDefinitionProvider", modelUri, new Position(13, 16));
+			const pythonPath: string = workspace.getConfiguration('inmanta').get<string>('pythonPath');
+			const inmantaVersion = await getInmantaVersion(pythonPath);
+			let expectedAttributeLocation;
+			if (inmantaVersion.major < 2020 || inmantaVersion.major === 2020 && inmantaVersion.minor <= 5) {
+				expectedAttributeLocation = new Range(new Position(2, 0), new Position(3, 0));
+			} else {
+				expectedAttributeLocation = new Range(new Position(2, 11), new Position(2, 15));
+			}
+			assert.strictEqual((attributeInSameFile as Location[]).length, 1);
+			assert.strictEqual(attributeInSameFile[0].uri.fsPath, modelUri.fsPath);
+			assert.deepStrictEqual(attributeInSameFile[0].range, expectedAttributeLocation, "Attribute location in the same file doesn't match");
+
+			const typeInDifferentFile = await commands.executeCommand("vscode.executeDefinitionProvider", modelUri, new Position(4, 18));
+			assert.strictEqual((typeInDifferentFile as Location[]).length, 1);
+			assert.strictEqual(typeInDifferentFile[0].uri.fsPath, path.resolve(libsPath, "testmodule", "model", "_init.cf"));
+			assert.deepStrictEqual(typeInDifferentFile[0].range, new Range(new Position(0, 8), new Position(0, 11)), "Attribute location in different file doesn't match");
+
+			if (inmantaVersion.major < 2020 || inmantaVersion.major === 2020 && inmantaVersion.minor <= 5) {
+				console.log("Plugin navigation not available on this Inmanta version");
+			} else {
+				const pluginLocation = await commands.executeCommand("vscode.executeDefinitionProvider", modelUri, new Position(17, 15));
+				assert.strictEqual((pluginLocation as Location[]).length, 1);
+				assert.strictEqual(pluginLocation[0].uri.fsPath, path.resolve(libsPath, "testmodule", "plugins", "__init__.py"));
+				assert.deepStrictEqual(pluginLocation[0].range, new Range(new Position(4, 0), new Position(5, 0)), "Plugin location doesn't match");
+			}
+			resolve();
+		});
+	}).timeout(0);
+	
+
+	after(async () => {
+		await Promise.all([
+			fs.writeFile(logPath, ""),
+		]);
+	});
+});